# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ClearPay Sales System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Lead {
  id: number;
  businessName: string;
  contactName: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  vertical?: string | null;
  /** @nullable */
  leadSource?: string | null;
  /** @nullable */
  existingPos?: string | null;
  /** @nullable */
  processor?: string | null;
  /** @nullable */
  estimatedMonthlyVolume?: number | null;
  leadScore: number;
  status: string;
  /** @nullable */
  assignedRepId?: number | null;
  /** @nullable */
  assignedRepName?: string | null;
  /** @nullable */
  lastContactDate?: string | null;
  /** @nullable */
  nextFollowUpDate?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateLeadBody {
  businessName: string;
  contactName: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  vertical?: string | null;
  /** @nullable */
  leadSource?: string | null;
  /** @nullable */
  existingPos?: string | null;
  /** @nullable */
  processor?: string | null;
  /** @nullable */
  estimatedMonthlyVolume?: number | null;
  status?: string;
  /** @nullable */
  assignedRepId?: number | null;
  /** @nullable */
  lastContactDate?: string | null;
  /** @nullable */
  nextFollowUpDate?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface UpdateLeadBody {
  businessName?: string;
  contactName?: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  vertical?: string | null;
  /** @nullable */
  leadSource?: string | null;
  /** @nullable */
  existingPos?: string | null;
  /** @nullable */
  processor?: string | null;
  /** @nullable */
  estimatedMonthlyVolume?: number | null;
  status?: string;
  /** @nullable */
  assignedRepId?: number | null;
  /** @nullable */
  lastContactDate?: string | null;
  /** @nullable */
  nextFollowUpDate?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface Note {
  id: number;
  leadId: number;
  /** @nullable */
  authorId?: number | null;
  /** @nullable */
  authorName?: string | null;
  content: string;
  createdAt: string;
}

export interface CreateNoteBody {
  content: string;
  /** @nullable */
  authorId?: number | null;
}

export interface Task {
  id: number;
  /** @nullable */
  leadId?: number | null;
  /** @nullable */
  leadName?: string | null;
  /** @nullable */
  assignedTo?: number | null;
  /** @nullable */
  assignedToName?: string | null;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  dueDate?: string | null;
  priority: string;
  completed: boolean;
  createdAt: string;
}

export interface CreateTaskBody {
  /** @nullable */
  leadId?: number | null;
  /** @nullable */
  assignedTo?: number | null;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  dueDate?: string | null;
  priority?: string;
}

export interface UpdateTaskBody {
  title?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  dueDate?: string | null;
  priority?: string;
  completed?: boolean;
  /** @nullable */
  assignedTo?: number | null;
}

export interface Rep {
  id: number;
  name: string;
  email: string;
  role: string;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}

export interface CreateRepBody {
  name: string;
  email: string;
  role?: string;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface UpdateRepBody {
  name?: string;
  email?: string;
  role?: string;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface RepStats {
  repId: number;
  totalLeads: number;
  closedWon: number;
  closedLost: number;
  inPipeline: number;
  conversionRate: number;
  totalVolume: number;
}

export interface OnboardingRecord {
  id: number;
  leadId: number;
  merchantName: string;
  applicationSubmitted: boolean;
  underwritingApproved: boolean;
  equipmentShipped: boolean;
  accountActivated: boolean;
  trainingCompleted: boolean;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateOnboardingBody {
  leadId: number;
  merchantName: string;
  /** @nullable */
  notes?: string | null;
}

export interface UpdateOnboardingBody {
  applicationSubmitted?: boolean;
  underwritingApproved?: boolean;
  equipmentShipped?: boolean;
  accountActivated?: boolean;
  trainingCompleted?: boolean;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  completedAt?: string | null;
}

export interface DashboardSummary {
  totalLeads: number;
  newLeadsThisMonth: number;
  closedWonThisMonth: number;
  totalPipelineValue: number;
  winRate: number;
  tasksDueToday: number;
  avgLeadScore: number;
}

export interface StatusCount {
  status: string;
  count: number;
  totalVolume: number;
}

export interface ActivityItem {
  id: string;
  type: string;
  description: string;
  /** @nullable */
  leadId?: number | null;
  /** @nullable */
  leadName?: string | null;
  /** @nullable */
  repName?: string | null;
  createdAt: string;
}

export interface RepLeaderboardEntry {
  repId: number;
  repName: string;
  totalLeads: number;
  closedWon: number;
  conversionRate: number;
  totalVolume: number;
}

export interface AiInsight {
  summary: string;
  recommendations: string[];
  scoreExplanation: string;
  nextBestAction: string;
}

export type ListLeadsParams = {
  status?: string;
  /**
   * @nullable
   */
  assignedRep?: number | null;
  search?: string;
  vertical?: string;
  leadSource?: string;
};

export type ListTasksParams = {
  /**
   * @nullable
   */
  assignedTo?: number | null;
  completed?: boolean;
  priority?: string;
  dueToday?: boolean;
};