# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ClearPay Sales System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List all leads
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: assignedRep
          schema:
            type: ["integer", "null"]
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: vertical
          schema:
            type: string
        - in: query
          name: leadSource
          schema:
            type: string
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a new lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadBody"
      responses:
        "201":
          description: Created lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get a lead by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeadBody"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete a lead
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /leads/{id}/score:
    post:
      operationId: recalculateLeadScore
      tags: [leads]
      summary: Recalculate the lead score
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}/notes:
    get:
      operationId: listLeadNotes
      tags: [notes]
      summary: List notes for a lead
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createLeadNote
      tags: [notes]
      summary: Create a note for a lead
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteBody"
      responses:
        "201":
          description: Created note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /leads/{id}/tasks:
    get:
      operationId: listLeadTasks
      tags: [tasks]
      summary: List tasks for a lead
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
  /leads/{id}/ai-insight:
    get:
      operationId: getLeadAiInsight
      tags: [ai]
      summary: Get AI insight for a lead
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AI insight
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiInsight"
...
  /onboarding:
    get:
      operationId: listOnboarding
      tags: [onboarding]
      summary: List onboarding records
      responses:
        "200":
          description: Onboarding records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OnboardingRecord"
    post:
      operationId: createOnboarding
      tags: [onboarding]
      summary: Create an onboarding record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOnboardingBody"
      responses:
        "201":
          description: Created record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingRecord"
  /onboarding/{id}:
    patch:
      operationId: updateOnboarding
      tags: [onboarding]
      summary: Update an onboarding record
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOnboardingBody"
      responses:
        "200":
          description: Updated record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingRecord"
...
components:
  schemas:
...
    Lead:
      type: object
      properties:
        id:
          type: integer
        businessName:
          type: string
        contactName:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        vertical:
          type: ["string", "null"]
        leadSource:
          type: ["string", "null"]
        existingPos:
          type: ["string", "null"]
        processor:
          type: ["string", "null"]
        estimatedMonthlyVolume:
          type: ["number", "null"]
        leadScore:
          type: integer
        status:
          type: string
        assignedRepId:
          type: ["integer", "null"]
        assignedRepName:
          type: ["string", "null"]
        lastContactDate:
          type: ["string", "null"]
        nextFollowUpDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - businessName
        - contactName
        - leadScore
        - status
        - createdAt
        - updatedAt
    CreateLeadBody:
      type: object
      properties:
        businessName:
          type: string
        contactName:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        vertical:
          type: ["string", "null"]
        leadSource:
          type: ["string", "null"]
        existingPos:
          type: ["string", "null"]
        processor:
          type: ["string", "null"]
        estimatedMonthlyVolume:
          type: ["number", "null"]
        status:
          type: string
        assignedRepId:
          type: ["integer", "null"]
        lastContactDate:
          type: ["string", "null"]
        nextFollowUpDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - businessName
        - contactName
    UpdateLeadBody:
      type: object
      properties:
        businessName:
          type: string
        contactName:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        vertical:
          type: ["string", "null"]
        leadSource:
          type: ["string", "null"]
        existingPos:
          type: ["string", "null"]
        processor:
          type: ["string", "null"]
        estimatedMonthlyVolume:
          type: ["number", "null"]
        status:
          type: string
        assignedRepId:
          type: ["integer", "null"]
        lastContactDate:
          type: ["string", "null"]
        nextFollowUpDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    Note:
      type: object
      properties:
        id:
          type: integer
        leadId:
          type: integer
        authorId:
          type: ["integer", "null"]
        authorName:
          type: ["string", "null"]
        content:
          type: string
        createdAt:
          type: string
      required:
        - id
        - leadId
        - content
        - createdAt
    CreateNoteBody:
      type: object
      properties:
        content:
          type: string
        authorId:
          type: ["integer", "null"]
      required:
        - content
    Task:
      type: object
      properties:
        id:
          type: integer
        leadId:
          type: ["integer", "null"]
        leadName:
          type: ["string", "null"]
        assignedTo:
          type: ["integer", "null"]
        assignedToName:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        priority:
          type: string
        completed:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - title
        - priority
        - completed
        - createdAt
...
    OnboardingRecord:
      type: object
      properties:
        id:
          type: integer
        leadId:
          type: integer
        merchantName:
          type: string
        applicationSubmitted:
          type: boolean
        underwritingApproved:
          type: boolean
        equipmentShipped:
          type: boolean
        accountActivated:
          type: boolean
        trainingCompleted:
          type: boolean
        notes:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - leadId
        - merchantName
        - applicationSubmitted
        - underwritingApproved
        - equipmentShipped
        - accountActivated
        - trainingCompleted
        - createdAt
        - updatedAt
    CreateOnboardingBody:
      type: object
      properties:
        leadId:
          type: integer
        merchantName:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - leadId
        - merchantName
    UpdateOnboardingBody:
      type: object
      properties:
        applicationSubmitted:
          type: boolean
        underwritingApproved:
          type: boolean
        equipmentShipped:
          type: boolean
        accountActivated:
          type: boolean
        trainingCompleted:
          type: boolean
        notes:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
...
    AiInsight:
      type: object
      properties:
        summary:
          type: string
        recommendations:
          type: array
          items:
            type: string
        scoreExplanation:
          type: string
        nextBestAction:
          type: string
      required:
        - summary
        - recommendations
        - scoreExplanation
        - nextBestAction