# lib/api-zod/src/generated/types/createNoteBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ClearPay Sales System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateNoteBody {
  content: string;
  /** @nullable */
  authorId?: number | null;
}