# lib/api-zod/src/generated/types/onboardingRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ClearPay Sales System API
 * OpenAPI spec version: 0.1.0
 */

export interface OnboardingRecord {
  id: number;
  leadId: number;
  merchantName: string;
  applicationSubmitted: boolean;
  underwritingApproved: boolean;
  equipmentShipped: boolean;
  accountActivated: boolean;
  trainingCompleted: boolean;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}